+++ src/commands.rs
@@ -1,4 +1,4 @@
-use crate::{data_types::Keygen, Keystore, KeystoreState};
+use crate::{data_types::Keygen, KeyRotationEvent, Keystore, KeystoreState};
 use bevy::{ecs::system::Command, prelude::*};
 use bevy_async_task::AsyncTask;
 
@@ -23,6 +23,7 @@ impl Command for StartKeyRotation {
         if keystore.access_token_valid_for() > crate::Duration::ZERO {
             let mut state = world.resource_mut::<NextState<KeystoreState>>();
             state.set(KeystoreState::Conformant);
+            world.send_event(KeyRotationEvent::Started(keystore.clone()));
         } else {
             warn!("auth provider authenticated, but returned an expired access token, remaining nonconformant");
         }
@@ -42,6 +43,7 @@ impl Command for StartKeyRotationWithKeystore {
         if keystore.access_token_valid_for() > crate::Duration::ZERO {
             let mut state = world.resource_mut::<NextState<KeystoreState>>();
             state.set(KeystoreState::Conformant);
+            world.send_event(KeyRotationEvent::Started(self.keystore.clone()));
         } else {
             warn!("started key rotation with an expired keystore, remaining nonconformant");
         }
@@ -73,6 +75,7 @@ impl Command for StopKeyRotation {
     fn apply(self, world: &mut bevy::prelude::World) {
         let mut state = world.resource_mut::<NextState<KeystoreState>>();
         state.set(KeystoreState::NonConformant);
+        world.send_event(KeyRotationEvent::Stopped);
         info!("stopping key rotation");
     }
 }