+++ src/lib.rs
@@ -10,7 +10,8 @@ pub use web_time::{Duration, Instant};
 
 pub use commands::{StartKeyRotationExt, StopKeyRotationExt};
 pub use data_types::{
-    AuthProvider, KeyRotationSettings, Keygen, Keystore, KeystoreState,
+    AuthProvider, KeyRotationEvent, KeyRotationSettings, Keygen, Keystore,
+    KeystoreState,
 };
 pub use error::TokenRotationError;
 pub use plugin::KeyRotationPlugin;